# src/frontend/csrs-portal/src/app/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CSRS API",
    "description": "BC Child Support Recalculation Service API",
    "version": "v1"
  },
  "paths": {
    "/api/account/genders": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Gets the valid gender values for an account.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/provinces": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Gets the valid province values for an account.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/identities": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Gets the valid identity values for an account.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/referrals": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Gets the valid referral values for an account.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/preferredcontactmethods": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Gets the preferred contact methods values for an account.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookupValue"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account/create": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Creates a new file.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user and file were created."
          },
          "400": {
            "description": "The request was not well formed.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "The request is not authorized. Ensure correct authentication header is present.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "There is already a draft file associated with the current user/account.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/checkcsrsaccount": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Check existence of CSRS account.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CSRSAccount"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CSRSAccount"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CSRSAccount"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The csrs account was found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CSRSPartyFileIds"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CSRSPartyFileIds"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CSRSPartyFileIds"
                }
              }
            }
          },
          "404": {
            "description": "The csrs account was not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "The request is not authorized. Ensure correct authentication header is present.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/account/updatecsrsaccount": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Update CSRS account.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CSRSAccountRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CSRSAccountRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CSRSAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The CSRS account was created."
          },
          "400": {
            "description": "The request was not well formed.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "The request is not authorized. Ensure correct authentication header is present.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Child": {
        "type": "object",
        "properties": {
          "childId": {
            "type": "string",
            "description": "The child id.",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Child's first name.",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "description": "Child's middle name.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Child's last name.",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Child's date of birth name.",
            "format": "date",
            "nullable": true
          },
          "childIsDependent": {
            "type": "string",
            "description": "Child is a dependent.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CourtLookupValue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique id of the value.",
            "nullable": true
          },
          "value": {
            "type": "string",
            "description": "The human readable value.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CSRSAccount": {
        "required": [
          "fileNumber",
          "referenceNumber"
        ],
        "type": "object",
        "properties": {
          "fileNumber": {
            "type": "string",
            "description": "The input FileNumber parameter to search CSRS account."
          },
          "referenceNumber": {
            "type": "string",
            "description": "The input Password parameter to search CSRS account."
          }
        },
        "additionalProperties": false
      },
      "CSRSAccountFile": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "nullable": true
          },
          "fileNumber": {
            "type": "string",
            "nullable": true
          },
          "partyEnrolled": {
            "type": "string",
            "nullable": true
          },
          "safetyAlertRecipient": {
            "type": "string",
            "nullable": true
          },
          "recipientSafetyConcernDescription": {
            "type": "string",
            "nullable": true
          },
          "safetyAlertPayor": {
            "type": "string",
            "nullable": true
          },
          "payorSafetyConcernDescription": {
            "type": "string",
            "nullable": true
          },
          "isFMEPFileActive": {
            "type": "string",
            "nullable": true
          },
          "fmepFileNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CSRSAccountRequest": {
        "required": [
          "csrsAccountFile",
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/Party"
          },
          "csrsAccountFile": {
            "$ref": "#/components/schemas/CSRSAccountFile"
          }
        },
        "additionalProperties": false
      },
      "CSRSPartyFileIds": {
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string",
            "description": "The output Party Id of CSRS account.",
            "nullable": true
          },
          "fileId": {
            "type": "string",
            "description": "The output File Id of CSRS account.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "File": {
        "required": [
          "usersRole"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/FileStatus"
          },
          "usersRole": {
            "$ref": "#/components/schemas/PartyRole"
          },
          "fileId": {
            "type": "string",
            "nullable": true
          },
          "fileNumber": {
            "type": "string",
            "nullable": true
          },
          "partyEnrolled": {
            "type": "string",
            "nullable": true
          },
          "courtFileType": {
            "$ref": "#/components/schemas/LookupValue"
          },
          "bcCourtLevel": {
            "$ref": "#/components/schemas/CourtLookupValue"
          },
          "bcCourtLocation": {
            "$ref": "#/components/schemas/CourtLookupValue"
          },
          "dateOfOrderOrWA": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "incomeOnOrder": {
            "type": "string",
            "nullable": true
          },
          "section7Expenses": {
            "type": "string",
            "nullable": true
          },
          "safetyAlertRecipient": {
            "type": "string",
            "nullable": true
          },
          "recipientSafetyConcernDescription": {
            "type": "string",
            "nullable": true
          },
          "safetyAlertPayor": {
            "type": "string",
            "nullable": true
          },
          "payorSafetyConcernDescription": {
            "type": "string",
            "nullable": true
          },
          "isFMEPFileActive": {
            "type": "string",
            "nullable": true
          },
          "fmepFileNumber": {
            "type": "string",
            "nullable": true
          },
          "recalculationOrderByCourt": {
            "type": "string",
            "nullable": true
          },
          "otherParty": {
            "$ref": "#/components/schemas/Party"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Child"
            },
            "description": "The children on this file.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FileStatus": {
        "enum": [
          "Unknown",
          "Draft",
          "Active"
        ],
        "type": "string"
      },
...
      "LookupValue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique id of the value.",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "description": "The human readable value.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NewFileRequest": {
        "required": [
          "file",
          "user"
        ],
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/Party"
          },
          "file": {
            "$ref": "#/components/schemas/File"
          }
        },
        "additionalProperties": false
      },
      "Party": {
        "required": [
          "dateOfBirth",
          "email",
          "firstName",
          "gender",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "partyId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string"
          },
          "preferredName": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "$ref": "#/components/schemas/LookupValue"
          },
          "addressStreet1": {
            "type": "string",
            "nullable": true
          },
          "addressStreet2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "province": {
            "$ref": "#/components/schemas/LookupValue"
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "homePhone": {
            "type": "string",
            "nullable": true
          },
          "workPhone": {
            "type": "string",
            "nullable": true
          },
          "cellPhone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string"
          },
          "optOutElectronicDocuments": {
            "type": "boolean",
            "nullable": true
          },
          "identity": {
            "$ref": "#/components/schemas/LookupValue"
          },
          "referral": {
            "$ref": "#/components/schemas/LookupValue"
          },
          "preferredContactMethod": {
            "$ref": "#/components/schemas/LookupValue"
          },
          "referenceNumber": {
            "type": "string",
            "nullable": true
          },
          "incomeAssistance": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PartyRole": {
        "enum": [
          "Unknown",
          "Recipient",
          "Payor"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}